M      test/server/j2se/com/sun/sgs/test/impl/service/nodemap/TestNodeMappingServiceImpl.java
 - testRoundRobinAutoMove did not account for task service calling assignNode on the 
   application identity.  This fixes an intermittent problem Seth ran into before
   integrating the multi-node task service.
 - Added test for getIdentities() on a node with no identities assigned.

M      src/server/j2se/com/sun/sgs/impl/service/nodemap/RoundRobinPolicy.java
 - clarify that watchdogService.getNode can return null in comment

M      src/server/j2se/com/sun/sgs/impl/service/nodemap/NodeMappingServiceImpl.java
 - throw UnknownIdentityException in getNode, if the watchdog service returns null
   for a node that is no longer in the data store
 - throw UnknownIdentityException in getIdentities if the watchdog service returns null.
   Remove incorrect code that prevented us from calling getIdentities on a node with
   no assignments.

M      src/server/j2se/com/sun/sgs/impl/service/nodemap/NodeMappingServerImpl.java
 - protect CheckTask.run method from null return from watchdog service.getNode
 - fix a couple of data races in mapToNewNode
    - The identity could have been removed before it was actually assigned in mapToNewNode
      (problem found by inspection).  If this happens, act as if it had never been assigned
      before, and assign it to the chosen new node.
    - The identity's old mappings could have been removed if two threads are trying to
      map to a new node at the same time.  This problem occurred because I was using data
      from outside a transaction to decide whether to remove bindings inside the transaction,
      and that data might have gone stale.  Fix is to check inside the transaction to see if
      the old name bindings need to be removed;  if they have been already, just return.
      Some other thread has already reassigned the identity.
      This is the problem I saw in the test TestKernelSimpleAppRestart.

+++ src/server/j2se/com/sun/sgs/impl/service/nodemap/NodeMappingServerImpl.java
@@ -465,10 +465,9 @@ public class NodeMappingServerImpl implements NodeMappingServer {
                 found = true;
                 long nodeId = idmo.getNodeId();
 
-                isAlive = watchdogService.getNode(nodeId).isAlive();
+                node = watchdogService.getNode(nodeId);
+                isAlive = (node != null && node.isAlive());
                 if (!isAlive) {
-                    // Caller can get result from getNodeId()
-                    node = watchdogService.getNode(nodeId);
                     return;
                 }
                 // The identity already has an assignment but we still 
@@ -628,7 +627,9 @@ public class NodeMappingServerImpl implements NodeMappingServer {
         boolean isDead() {
             return dead;
         }
-        /** Returns the node id for the node the identity was removed from. */
+        /** Returns the node the identity was removed from, which can be
+         *  null if the node has failed and been removed from the data store.
+         */
         Node getNode() {
             return node;
         } 
@@ -800,24 +801,39 @@ public class NodeMappingServerImpl implements NodeMappingServer {
                 public void run() {                   
                     // First, we clean up any old mappings.
                     if (oldNode != null) {
-                        // Find the old IdentityMO, with the old node info.
-                        IdentityMO oldidmo = 
-                            dataService.getServiceBinding(idkey, 
-                                                          IdentityMO.class);
-                        //Remove the old node->id key.
-                        dataService.removeServiceBinding(oldNodeKey);
+                        try {
+                            // Find the old IdentityMO, with the old node info.
+                            IdentityMO oldidmo = 
+                                dataService.getServiceBinding(idkey, 
+                                                              IdentityMO.class);
 
-                        // Remove the old status information.  We don't 
-                        // retain any info about the old node's status.
-                        Iterator<String> iter =
-                            BoundNamesUtil.getServiceBoundNamesIterator(
-                                dataService, oldStatusKey);
-                        while (iter.hasNext()) {
-                            iter.next();
-                            iter.remove();
+                            // Check once more for the assigned node - someone
+                            // else could have mapped it before we got here.
+                            // If so, just return.
+                            if (oldidmo.getNodeId() != oldNode.getId()) {
+                                return;
+                            }
+
+                            //Remove the old node->id key.
+                            dataService.removeServiceBinding(oldNodeKey);
+
+                            // Remove the old status information.  We don't 
+                            // retain any info about the old node's status.
+                            Iterator<String> iter =
+                                BoundNamesUtil.getServiceBoundNamesIterator(
+                                    dataService, oldStatusKey);
+                            while (iter.hasNext()) {
+                                iter.next();
+                                iter.remove();
+                            }
+                            // Remove the old IdentityMO with the old node info.
+                            dataService.removeObject(oldidmo);
+                        } catch (NameNotBoundException e) {
+                            // The identity was removed before we could
+                            // reassign it to a new node.
+                            // Simply make the new assignment, as if oldNode
+                            // was null to begin with.
                         }
-                        // Remove the old IdentityMO with the old node info.
-                        dataService.removeObject(oldidmo);
                     }
                     // Add (or update) the id->node mapping. 
                     dataService.setServiceBinding(idkey, newidmo);
@@ -872,7 +888,8 @@ public class NodeMappingServerImpl implements NodeMappingServer {
                              
         /**             
          * Returns the node found by the watchdog service, or null if
-         * this task has not run.
+         * this task has not run or the node has failed and been removed
+         * from the data store.
          */
         public Node getNode() {
             return node;
